backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic[email]==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1